[opentitanlib] move openocd/elf_debugger code to dedicated `debug` module
