+++ sw/host/opentitanlib/BUILD
@@ -82,6 +82,9 @@ rust_library(
         "src/crypto/rsa.rs",
         "src/crypto/sha256.rs",
         "src/crypto/spx.rs",
+        "src/debug/elf_debugger.rs",
+        "src/debug/mod.rs",
+        "src/debug/openocd.rs",
         "src/dif/aon_timer.rs",
         "src/dif/clkmgr.rs",
         "src/dif/lc_ctrl.rs",
@@ -122,7 +125,6 @@ rust_library(
         "src/spiflash/sfdp.rs",
         "src/test_utils/bootstrap.rs",
         "src/test_utils/e2e_command.rs",
-        "src/test_utils/elf_debugger.rs",
         "src/test_utils/epmp.rs",
         "src/test_utils/extclk.rs",
         "src/test_utils/gpio.rs",
@@ -198,7 +200,6 @@ rust_library(
         "src/util/file.rs",
         "src/util/mod.rs",
         "src/util/num_de.rs",
-        "src/util/openocd.rs",
         "src/util/parse_int.rs",
         "src/util/present.rs",
         "src/util/printer.rs",
