+++ sw/host/opentitanlib/src/debug/elf_debugger.rs
