+++ sw/host/opentitanlib/src/debug/mod.rs
@@ -1,0 +1,8 @@
+// Copyright lowRISC contributors.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Debugging-related facilities.
+
+pub mod elf_debugger;
+pub mod openocd;
