+++ sw/host/opentitanlib/src/io/jtag.rs
@@ -11,9 +11,9 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 use crate::app::TransportWrapper;
+use crate::debug::openocd::OpenOcd;
 use crate::dif::lc_ctrl::LcCtrlReg;
 use crate::impl_serializable_error;
-use crate::util::openocd::OpenOcd;
 
 #[derive(Debug, Args, Clone)]
 pub struct JtagParams {
