+++ sw/host/opentitanlib/src/test_utils/mod.rs
@@ -4,7 +4,6 @@
 
 pub mod bootstrap;
 pub mod e2e_command;
-pub mod elf_debugger;
 pub mod epmp;
 pub mod extclk;
 pub mod gpio;
